+++ packages/builder/builder-rspack-provider/src/types/config/html.ts
@@ -1,0 +1,5 @@
+import type { SharedHtmlConfig } from '@modern-js/builder-shared';
+
+export type HtmlConfig = SharedHtmlConfig;
+
+export type NormalizedHtmlConfig = HtmlConfig;