+++ packages/builder/builder-rspack-provider/src/types/config/security.ts
@@ -1,0 +1,5 @@
+// TODO
+// eslint-disable-next-line @typescript-eslint/no-empty-interface
+export interface SecurityConfig {}
+
+export type NormalizedSecurityConfig = Required<SecurityConfig>;