+++ packages/builder/builder-webpack-provider/src/config/defaults.ts
@@ -1,17 +1,10 @@
 import {
   extendsType,
-  DEFAULT_PORT,
-  ROOT_DIST_DIR,
-  HTML_DIST_DIR,
-  JS_DIST_DIR,
-  CSS_DIST_DIR,
-  SVG_DIST_DIR,
-  FONT_DIST_DIR,
-  IMAGE_DIST_DIR,
-  MEDIA_DIST_DIR,
-  SERVER_DIST_DIR,
   mergeBuilderConfig,
-  DEFAULT_DATA_URL_SIZE,
+  defaultDevConfig,
+  defaultOutputConfig,
+  defaultHtmlConfig,
+  defaultSourceConfig,
 } from '@modern-js/builder-shared';
 import type { BuilderConfig } from '../types';
 
@@ -19,18 +12,8 @@ const defineDefaultConfig = extendsType<BuilderConfig>();
 
 export const createDefaultConfig = () =>
   defineDefaultConfig({
-    dev: {
-      hmr: true,
-      https: false,
-      port: DEFAULT_PORT,
-      assetPrefix: '/',
-      startUrl: false,
-      progressBar: true,
-    },
-    html: {
-      crossorigin: false,
-      disableHtmlFolder: false,
-    },
+    dev: defaultDevConfig,
+    html: defaultHtmlConfig,
     tools: {
       cssExtract: {
         loaderOptions: {},
@@ -38,49 +21,8 @@ export const createDefaultConfig = () =>
       },
       tsChecker: true,
     },
-    source: {
-      define: {},
-      alias: {},
-      preEntry: [],
-      globalVars: {},
-      compileJsDataURI: true,
-    },
-    output: {
-      distPath: {
-        root: ROOT_DIST_DIR,
-        js: JS_DIST_DIR,
-        css: CSS_DIST_DIR,
-        svg: SVG_DIST_DIR,
-        font: FONT_DIST_DIR,
-        html: HTML_DIST_DIR,
-        image: IMAGE_DIST_DIR,
-        media: MEDIA_DIST_DIR,
-        server: SERVER_DIST_DIR,
-      },
-      filename: {},
-      charset: 'ascii',
-      polyfill: 'entry',
-      dataUriLimit: {
-        svg: DEFAULT_DATA_URL_SIZE,
-        font: DEFAULT_DATA_URL_SIZE,
-        image: DEFAULT_DATA_URL_SIZE,
-        media: DEFAULT_DATA_URL_SIZE,
-      },
-      legalComments: 'linked',
-      cleanDistPath: true,
-      svgDefaultExport: 'url',
-      disableMinimize: false,
-      disableSourceMap: false,
-      disableFilenameHash: false,
-      disableCssModuleExtension: false,
-      disableInlineRuntimeChunk: false,
-      enableAssetFallback: false,
-      enableAssetManifest: false,
-      enableLatestDecorators: false,
-      enableCssModuleTSDeclaration: false,
-      enableInlineScripts: false,
-      enableInlineStyles: false,
-    },
+    source: defaultSourceConfig,
+    output: defaultOutputConfig,
     security: { sri: false },
     experiments: {},
     performance: {