+++ packages/cli/plugin-lint/tests/index.test.ts
@@ -1,0 +1,7 @@
+import plugin from '../src';
+
+describe('plugin-lint', () => {
+  it('default', () => {
+    expect(plugin).toBeDefined();
+  });
+});