+++ packages/solutions/module-tools-v2/src/plugins.ts
@@ -1,6 +1,6 @@
 import type { CliPlugin } from '@modern-js/core';
 import ChangesetPlugin from '@modern-js/plugin-changeset';
-import LintPlugin from '@modern-js/plugin-jarvis';
+import LintPlugin from '@modern-js/plugin-lint';
 
 export const getPlugins = (runningCmd: string) => {
   let plugins: CliPlugin[] = [];