+++ tests/integration/server-prod/test/index.test.js
@@ -42,14 +42,15 @@ describe('test basic usage', () => {
     expect(fs.readFileSync(mainEntry, 'utf-8')).toMatch(
       '<link rel="icon" href="/favicon1.ico">',
     );
+    const mediaPath = path.join('static', 'media', 'icon.png');
     expect(fs.readFileSync(mainEntry, 'utf-8')).toMatch(
-      '<link rel="apple-touch-icon" sizes="180*180" href="/static/media/icon.png">',
+      `<link rel="apple-touch-icon" sizes="180*180" href="/${mediaPath}">`,
     );
     expect(fs.readFileSync(activityEntry, 'utf-8')).toMatch(
       '<link rel="icon" href="/favicon.ico">',
     );
     expect(fs.readFileSync(activityEntry, 'utf-8')).toMatch(
-      '<link rel="apple-touch-icon" sizes="180*180" href="/static/media/icon.png">',
+      `<link rel="apple-touch-icon" sizes="180*180" href="/${mediaPath}">`,
     );
   });